requirements.txt
# Deep Learning Framework
torch
torchvision

# HuggingFace Libraries
datasets
transformers

# Data Science & Visualization
matplotlib
scipy
numpy
pandas

# Configuration & Environment
python-dotenv
pyyaml

# Image Processing
Pillow

# Dataset Downloads (optional)
kaggle
tqdm
ipywidgets
jupyter
ipykernal
